// src/app/shared/model/order-detail.model.ts
export class OrderDetail {
  ExportStoreDetailID: number;
  ExportID: number;
  ProductID: number;
  SpecID1: number;
  SpecID2: number;
  SpecID3: number;
  ProductionID: number;
  Quantity: number;
  Amount: number;
  Total: number;
  Description: string;
  NameProduct: string;
  NameProduction: string;
  SpecName1: string;
  SpecName2: string;
  SpecName3: string;

  constructor() {
  }
}